+++ git_machete/git_operations.py
@@ -180,6 +180,16 @@ class BranchPair(NamedTuple):
 
 HEAD = AnyRevision.of("HEAD")
 
+# For simplicity, explicitly suppress showing GPG signatures in `git log`
+# operations to simplify log parsing. This option must be passed as a
+# configuration parameter because the `git log` command's `--no-show-signature`
+# flag does not exist prior to `git` version 2.10.0; `git` does not emit an
+# error if it is passed via the `-c` flag a configuration setting that does not
+# exist, so compatibility with `git` versions earlier than version 2.10.0 is
+# preserved, even though the `log.showSignature` setting also does not exist
+# prior to version 2.10.0. Fixes a bug documented in GitHub issue #1286.
+GIT_EXEC = ("git", "-c", "log.showSignature=false")
+
 
 class GitContext:
 
@@ -224,7 +234,7 @@ class GitContext:
         self.__short_commit_hash_by_revision_cached = {}
 
     def _run_git(self, git_cmd: str, *args: str, flush_caches: bool, allow_non_zero: bool = False) -> int:
-        exit_code = utils.run_cmd("git", git_cmd, *args)
+        exit_code = utils.run_cmd(*GIT_EXEC, git_cmd, *args)
         if flush_caches:
             self.flush_caches()
         if not allow_non_zero and exit_code != 0:
@@ -234,7 +244,7 @@ class GitContext:
 
     def _popen_git(self, git_cmd: str, *args: str,
                    allow_non_zero: bool = False, env: Optional[Dict[str, str]] = None, input: Optional[str] = None) -> CommandResult:
-        exit_code, stdout, stderr = utils.popen_cmd("git", git_cmd, *args, env=env, input=input)
+        exit_code, stdout, stderr = utils.popen_cmd(*GIT_EXEC, git_cmd, *args, env=env, input=input)
         if not allow_non_zero and exit_code != 0:
             exit_code_msg: str = fmt(f"`{utils.get_cmd_shell_repr('git', git_cmd, *args, env=env)}` returned {exit_code}\n")
             stdout_msg: str = f"\n{utils.bold('stdout')}:\n{utils.dim(stdout)}" if stdout else ""